Punto 4 y correccion de errores

+++ src/main/java/ar/edu/unlam/tallerweb1/modelo/Ubicacion.java
@@ -10,8 +10,8 @@ public class Ubicacion {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-	private Integer latitud;
-	private Integer longitud;
+	private Double latitud;
+	private Double longitud;
 	
 	public Long getId() {
 		return id;
@@ -19,16 +19,16 @@ public class Ubicacion {
 	public void setId(Long id) {
 		this.id = id;
 	}
-	public Integer getLatitud() {
+	public Double getLatitud() {
 		return latitud;
 	}
-	public void setLatitud(Integer latitud) {
+	public void setLatitud(Double latitud) {
 		this.latitud = latitud;
 	}
-	public Integer getLongitud() {
+	public Double getLongitud() {
 		return longitud;
 	}
-	public void setLongitud(Integer longitud) {
+	public void setLongitud(Double longitud) {
 		this.longitud = longitud;
 	}
